pre - config re-structering
